requirements.txt
pandas
bs4
requests
vertexai
langchain
huggingface-hub
langchain_google_vertexai
langchain-community
tf-keras
pypdf
chromadb
google-cloud-aiplatform
langchain_huggingface
streamlit
streamlit-chat
crawl4ai
nest_asyncio
langchain
langchain-community
langchain-core

langchain-ollama
faiss-cpu
unstructured
markdown
python-dotenv
playwright
numpy
scipy
transformers 
sentence-transformers
